+++ R/format_mr_results2.R
@@ -146,32 +146,9 @@ combine_all_mrresults <- function(
   het <- het[, c("id.exposure", "id.outcome", "method", "Q", "Q_df", "Q_pval")]
 
   # Convert all factors to character
-  # lapply(names(Res), FUN=function(x) class(Res[,x]))
-  Class <- unlist(lapply(names(res), FUN = function(x) class(res[, x])))
-  if (any(Class == "factor")) {
-    Pos <- which(unlist(lapply(names(res), FUN = function(x) class(res[, x]))) == "factor")
-    for (i in seq_along(Pos)) {
-      res[, Pos[i]] <- as.character(res[, Pos[i]])
-    }
-  }
-
-  # lapply(names(Het), FUN=function(x) class(Het[,x]))
-  Class <- unlist(lapply(names(het), FUN = function(x) class(het[, x])))
-  if (any(Class == "factor")) {
-    Pos <- which(unlist(lapply(names(het), FUN = function(x) class(het[, x]))) == "factor")
-    for (i in seq_along(Pos)) {
-      het[, Pos[i]] <- as.character(het[, Pos[i]])
-    }
-  }
-
-  # lapply(names(Sin), FUN=function(x) class(Sin[,x]))
-  Class <- unlist(lapply(names(sin), FUN = function(x) class(sin[, x])))
-  if (any(Class == "factor")) {
-    Pos <- which(unlist(lapply(names(sin), FUN = function(x) class(sin[, x]))) == "factor")
-    for (i in seq_along(Pos)) {
-      sin[, Pos[i]] <- as.character(sin[, Pos[i]])
-    }
-  }
+  res[] <- lapply(res, function(x) if (is.factor(x)) as.character(x) else x)
+  het[] <- lapply(het, function(x) if (is.factor(x)) as.character(x) else x)
+  sin[] <- lapply(sin, function(x) if (is.factor(x)) as.character(x) else x)
 
   sin <- sin[grep("[:0-9:]", sin$SNP), ]
   sin$method <- "Wald ratio"
@@ -185,10 +162,11 @@ combine_all_mrresults <- function(
   names(sin)[names(sin) == "method"] <- "Method"
 
   res <- merge(res, het, by = c("id.outcome", "id.exposure", "Method"), all.x = TRUE)
-  res <- plyr::rbind.fill(
-    res,
-    sin[, c("exposure", "outcome", "id.exposure", "id.outcome", "SNP", "b", "se", "pval", "Method")]
+  res <- data.table::rbindlist(
+    list(res, sin[, c("exposure", "outcome", "id.exposure", "id.outcome", "SNP", "b", "se", "pval", "Method")]),
+    fill = TRUE, use.names = TRUE
   )
+  data.table::setDF(res)
 
   if (ao_slc) {
     ao <- available_outcomes()